finalize

+++ FinalProject/Controllers/api/ManagementController.cs
@@ -1,49 +1,71 @@
 ﻿using AutoMapper;
+using FinalProject.Authentication;
 using FinalProject.DTOs;
 using FinalProject.Models;
-using System;
-using System.Data.Entity;
 using System.Collections.Generic;
+using System.Data.Entity;
 using System.Linq;
-using System.Net;
-using System.Net.Http;
 using System.Web.Http;
-using System.IO;
 
 namespace FinalProject.Controllers.api
 {
     //Add Authrization
+    [ManagerAuthentication]
     public class ManagementController : ApiController
     {
         private MyAppContext db = new MyAppContext();
-        // GET api/<controller>
-        #region Sign Up Request
 
+        #region Sign Up Request
         [HttpGet]
-        [Route("api/Management/SignUpRequests")]
-        public IHttpActionResult SignUpRequests()
+        [Route("api/Management/PharmaciesSignUpRequests")]
+        public IHttpActionResult PharmaciesSignUpRequests()
         {
-            var signUpRequests = db.Clinics.Include(c => c.ForUser).Include(c => c.ClinicType).Where(c => c.IsActiveClinic == false).Select(c => c).ToList();
-            if (signUpRequests.Count == 0)
+            var signUpRequestsClinics = db.Clinics.Include(c => c.ForUser).Include(c => c.ClinicType).Where(c => c.IsActiveClinic == false).Select(c => c);
+            var signUpRequestsPharamacies = db.Pharmacies.Include(c => c.ForUser).Where(c => c.IsActivePharmacy == false).Select(c => c);
+
+            if (signUpRequestsClinics.Count() == 0 && signUpRequestsPharamacies.Count() == 0)
             {
-                return Ok(signUpRequests);
+                return Ok(new List<SignUpRequestDTO>());//Empty List
             }
 
-
-            var result = signUpRequests.Select(r => new
+            var pharmacies = signUpRequestsPharamacies.Select(pharmacy => new SignUpRequestDTO
             {
-                clinicId = r.ClinicId,
-                clinicName = r.ClinicName,
-                clinicType = r.ClinicType.ClinicTypeName,
-                doctorName = $"{r.ForUser.FirstName} {r.ForUser.LastName}",
-                certificate = r.Certificate
+                Id = pharmacy.PharmacyId,
+                Name = pharmacy.PharmacyName,
+                Type = "Phamracy",
+                DoctorName = pharmacy.ForUser.FirstName + " " + pharmacy.ForUser.LastName,
+                Certificate = pharmacy.Certificate
             });
 
-            return Ok(result);
+            return Ok(pharmacies.ToList());
+        }
+
+        [HttpGet]
+        [Route("api/Management/ClinicsSignUpRequests")]
+        public IHttpActionResult ClinicsSignUpRequests()
+        {
+            var signUpRequestsClinics = db.Clinics.Include(c => c.ForUser).Include(c => c.ClinicType).Where(c => c.IsActiveClinic == false).Select(c => c);
+            var signUpRequestsPharamacies = db.Pharmacies.Include(c => c.ForUser).Where(c => c.IsActivePharmacy == false).Select(c => c);
+
+            if (signUpRequestsClinics.Count() == 0 && signUpRequestsPharamacies.Count() == 0)
+            {
+                return Ok(new List<SignUpRequestDTO>());//Empty List
+            }
+
+            var clinics = signUpRequestsClinics.Select(clinic => new SignUpRequestDTO
+            {
+                Id = clinic.ClinicId,
+                Name = clinic.ClinicName,
+                Type = clinic.ClinicType.ClinicTypeName,
+                DoctorName = clinic.ForUser.FirstName + " " + clinic.ForUser.LastName,
+                Certificate = clinic.Certificate
+            });
+
+            return Ok(clinics.ToList());
         }
 
         [HttpPost]
-        [Route("api/Management/Clinics/Accept/{id}")]
+        [Route("api/Management/Clinics/AcceptClinic/{id}")]
         public IHttpActionResult AcceptClinic(int id)
         {
             var clinic = db.Clinics.FirstOrDefault(c => c.ClinicId == id && c.IsActiveClinic == false);
@@ -56,7 +78,7 @@ namespace FinalProject.Controllers.api
         }
 
         [HttpDelete]
-        [Route("api/Management/Clinics/Refuse/{id}")]
+        [Route("api/Management/Clinics/RefuseClinic/{id}")]
         public IHttpActionResult RefuseClinic(int id)
         {
             var clinic = db.Clinics.FirstOrDefault(c => c.ClinicId == id && c.IsActiveClinic == false);
@@ -67,9 +89,39 @@ namespace FinalProject.Controllers.api
             //Send email contains why
             var location = db.Locations.FirstOrDefault(l => l.ClinicId == clinic.ClinicId);
             var schedule = db.Schedules.FirstOrDefault(s => s.ClinicId == clinic.ClinicId);
-            db.Locations.Remove(location);
-            db.Schedules.Remove(schedule);
+            if (location != null)
+                db.Locations.Remove(location);
+            if (schedule != null)
+                db.Schedules.Remove(schedule);
             db.Clinics.Remove(clinic);
+            db.SaveChanges();
+            return Ok();
+        }
+
+        [HttpPost]
+        [Route("api/Management/Clinics/AcceptPharmacy/{id}")]
+        public IHttpActionResult AcceptPharmacy(int id)
+        {
+            var pharmacy = db.Pharmacies.FirstOrDefault(c => c.PharmacyId == id && c.IsActivePharmacy == false);
+            if (pharmacy == null)
+                return NotFound();
+
+            pharmacy.IsActivePharmacy = true;
+            db.SaveChanges();
+            return Ok();
+        }
+
+        [HttpDelete]
+        [Route("api/Management/Clinics/RefusePharmacy/{id}")]
+        public IHttpActionResult RefusePharmacy(int id)
+        {
+            var pharamcy = db.Pharmacies.FirstOrDefault(p => p.PharmacyId == id && p.IsActivePharmacy == false);
+            if (pharamcy == null)
+                return NotFound();
+
+            //ToDo
+            //Send email contains why
+            db.Pharmacies.Remove(pharamcy);
             db.SaveChanges();
             return Ok();
         }
@@ -82,7 +134,7 @@ namespace FinalProject.Controllers.api
         [Route("api/Management/Clinics")]
         public IHttpActionResult GetClinic()
         {
-            var clinics = db.Clinics.Include(c => c.ForUser).Include(c=>c.ClinicType).ToList();
+            var clinics = db.Clinics.Include(c => c.ForUser).Include(c => c.ClinicType).ToList();
             if (clinics.Count == 0)
                 return Ok(clinics);//Retune empty list
 
@@ -102,15 +154,91 @@ namespace FinalProject.Controllers.api
 
         #endregion
 
+        #region Pharmacies
+
+        [HttpGet]
+        [Route("api/Management/Pharmacies")]
+        public IHttpActionResult GetPharmacies()
+        {
+            var pharmacies = db.Pharmacies.Include(c => c.ForUser).ToList();
+            if (pharmacies.Count == 0)
+                return Ok(pharmacies);//Retune empty list
+
+            var result = pharmacies.Select(phramacy => new
+            {
+                phramacyId = phramacy.PharmacyId,
+                phramacyName = phramacy.PharmacyName,
+                doctorName = $"{phramacy.ForUser.FirstName} {phramacy.ForUser.LastName}",
+                certificate = phramacy.Certificate,
+                isActivePharmacy = phramacy.IsActivePharmacy,
+                doctorId = phramacy.UserId
+            });
+
+            return Ok(result);
+        }
+
+        #endregion
+
+        #region ClinicTypes
+        [HttpDelete]
+        [Route("api/Management/ClinicTypes/{id}")]
+        public IHttpActionResult GetClinicTypes(int id)
+        {
+            var clinicType = db.ClinicTypes.FirstOrDefault(c => c.ClinicTypeId == id);
+            if (clinicType == null)
+                return NotFound();
+
+            var clinics = db.Clinics.Where(c => c.ClinicTypeId == id);
+            if (clinics.Count() != 0)
+                return BadRequest();
+
+            db.ClinicTypes.Remove(clinicType);
+            db.SaveChanges();
+            return Ok(Mapper.Map<ClinicType, ClinicTypeDTO>(clinicType));
+        }
+
+        [HttpGet]
+        [Route("api/Management/ClinicTypes")]
+        public IHttpActionResult GetClinicTypes()
+        {
+            return Ok(db.ClinicTypes.Select(Mapper.Map<ClinicType, ClinicTypeDTO>).ToList());
+        }
+        #endregion
+
+        #region MedicineTypes
+        [HttpDelete]
+        [Route("api/Management/MedicineTypes/{id}")]
+        public IHttpActionResult GetMedicineTypes(int id)
+        {
+            var MedicineType = db.MedicineTypes.FirstOrDefault(c => c.MedicineTypeId == id);
+            if (MedicineType == null)
+                return NotFound();
+
+            var Medicines = db.Medicines.Where(c => c.MedicineTypeId == id);
+            if (Medicines.Count() != 0)
+                return BadRequest();
+
+            db.MedicineTypes.Remove(MedicineType);
+            db.SaveChanges();
+            return Ok(Mapper.Map<MedicineType, MedicineTypeDTO>(MedicineType));
+        }
+
+        [HttpGet]
+        [Route("api/Management/MedicineTypes")]
+        public IHttpActionResult GetMedicineTypes()
+        {
+            return Ok(db.MedicineTypes.Select(Mapper.Map<MedicineType, MedicineTypeDTO>).ToList());
+        }
+        #endregion
+
         #region Cities
 
         [HttpGet]
         [Route("api/Management/Cities")]
         public IHttpActionResult GetCities()
         {
-            return Ok(db.Cities.Select(Mapper.Map<City,CityDTO>).ToList());
+            return Ok(db.Cities.Select(Mapper.Map<City, CityDTO>).ToList());
         }
-
 
         // GET api/<controller>/5
         [HttpGet]
@@ -121,7 +249,24 @@ namespace FinalProject.Controllers.api
             if (city == null)
                 return NotFound();
 
-            return Ok( Mapper.Map<City,CityDTO>(city));
+            return Ok(Mapper.Map<City, CityDTO>(city));
+        }
+
+        [HttpDelete]
+        [Route("api/Management/Cities/{id}")]
+        public IHttpActionResult DeleteCity(int id)
+        {
+            var city = db.Cities.FirstOrDefault(c => c.CityId == id);
+            if (city == null)
+                return NotFound();
+
+            var locations = db.Locations.Where(l => l.CityId == id);
+            if (locations.Count() != 0)
+                return BadRequest();
+
+            db.Cities.Remove(city);
+            db.SaveChanges();
+            return Ok(Mapper.Map<City, CityDTO>(city));
         }
 
         #endregion
